Update the Texas Hold'em benchmark.
Based on profiling, we can play a full round of Texas Hold'em in
approximately 8.4 micro-seconds (with the vast majority of the time spent in
computing the winning hand, indicating a potential direction for future
optimizations).

Note: the previous implementation of the benchmark was dominated by
overheads in the random play. This shows that the implementation of the game
engine itself is extremely efficient, and it's up to users to choose which
level of abstraction they would like to program against.
PiperOrigin-RevId: 266123645