Optimize the performance of Texas Hold'em.
This change improves the performance of the Texas Hold'em engine by almost 2x by using more efficient data structures when computing the hand rank of a Texas Hold'em hand. In general, this change swaps dictionaries for arrays, and attempts to avoid a few branches. Before, the Texas Hold'em HandRank initializer comprised over 50% of the CPU cycles. Now, the same initializer is about 18%.

computeWinner has shrunk from ~74% down to ~44%, so while it's still a relatively hot-spot on the profile, and the bulk of the time continues to be in the initializer, so future optimizations could be good.

PiperOrigin-RevId: 266146520