Do not initialize the cumulative values as the case sum to 0 is taken care of.
Both in RegretMatching and in the computation of the average policy, we do take care of the case when the sum of regrets or cumulative policies is zero.

Thus, the initial non zero cumulative values, which was necessary to prevent a division by zero, is no longer necessary.

Also, we do not set the cumulative policy to a non-zero value.

PiperOrigin-RevId: 270695400
Change-Id: Iba9dbe14b193d9e809a7c65bf9cd6102fbac3260