Clean up the OpenSpiel bot interface.
One may assume bots with different capabilities, from a simple "Step" and "Restart" interface, to a more complex bot, aware of it's own stochasticity and providing its full policy at each state.

I am using a single class with optional methods, and not two classes which would be more C++-friendly, because it's exposed in Python using pybind11 and having a single class is way easier to wrap.

PiperOrigin-RevId: 282344694
Change-Id: I82ab36ba977face3a07caeb1d289d89c188d511f