Make it possible to have "overridable" bots.
For some bots, it can be useful to set the state to a particular configuration.
It may be easier to implement ForceAction(state, action) than RestartAt(state), as with the later, we may need to replay some part of the history (e.g. to distinguish the actions that correspond to the player the bot is playing, one may need to replay all the actions to get all the intermediate states, in order to get the states current_player. This is because "History()" just stores the actions and not who played them).

Another (non contradictory) path, would be to save the player along side the past actions, so we know who played what.

It also changes the constructor to be a default one, with additional methods to call to notify that some features are available. This is to prevent having to update all usage every-time.

PiperOrigin-RevId: 282900739
Change-Id: I5d3a67eaa515b64a589e33a42abdebd76ca7c52e