Implement dynamic routing game and its mean field limit game
The dynamic routing game is very similar to the game described in:
- Multi-Agent Reinforcement Learning for Dynamic Routing Games: A Unified Paradigm by Z. Shoua, X. Di, 2020
We consider:
- a network given by the class Network.
- a list of vehicles given by the class Vehicle.
Each vehicle has an origin road section and would like to reach a destination road section in the network. At every time step the vehicle might exit its current road section, based on a probability given as a function of the volume of vehicle on its road section. If the vehicle exit the road section then it can choose on which successor road section it would like to go. When the vehicle reaches its destination, it gets the current time as its cost. Therefore each vehicle would like to minimize their travel time to reach their destination. If the vehicle does not reached its destination by the end of the game, then its cost is the number of time steps + 1.

The mean field routing game is a variant of the ones described by:
- A mean field route choice game by R. Salhab, J. Le Ny and R. P. Malhamé, 2018 IEEE CDC.
- Existence and uniqueness result for mean field games with congestion effect on graphs, O. Gueant, Applied Mathematics & Optimization, 2015
- Dynamic driving and routing games for autonomous vehicles on networks: A mean field game approach, K. Huang, X. Chen, X. Di and Q. Du, TRB part C, 2021
It is the extension of the dynamic routing game python_dynamic_routing_game. The list of vehicles decribing the N player of the dynamic routing game is replaced by a list of OriginDestinationDemand. One OriginDestinationDemand corresponds to one population of vehicles (with the same origin, destination and departure time).

This game is a variant of the mean field route choice game as the vehicle movement depends on the current network congestion. In the mean field route choice game, the number of time step to reach the destination is constant and does not depend on the network congestion, neither of the vehicle cost function. In the dynamic driving and routing games the vehicle choose its speed to travel on each link in order to minimize its cost function. Therefore the congestion is encoded in the cost function.

They are remaining to dos in the tests, they can be done in parallel to the code review. They are two remaining to dos in the code: the implementation of departure time (which can be done after the code is merged, or before if this seems useful) and implementation of deterministic travel time (which will not be done).
Let me know if there is anything to change.