Add policy value distribution computation
Currently the policy value modules only computes the expected return of a player in the mean field game. However, one can also want to get the distribution of the returns of a player and not only the expected return. This should enable doing chance programming on the distribution of returns (for example minimize expected return under the constraints that at least 90% of the returns should be below a given value).