Convert quoridor movement action IDs to relative
The original implementation of Quoridor used absolute position numbering
for pawn moves, so moving to b3 was always the same action ID, regardless
of where the pawn was. This commit changes to using relative action IDs,
so moving directly north is always the same action ID, regardless of
what square that is moving to.