Replace all uses of kNumCells in UltimateTTT
The use of kNumCells as a constant forces a TTT configuration.
Replace these hardcoded uses by dynamic values to allow other
TTT sizes.

As a side effect, UltimateTTTState is now a friend of TTTState,
since the ultimate TTT is a TTT. This makes it easier to share
members between these highly related classes.

Signed-off-by: odanrc <odanrc@yahoo.com.br>