Check legal actions for a board in TTT
Create a function that returns what are the available legal
actions for a given board. Notice that this is likely a
workaround, and it is possible that what we need is to keep
track of the view of each player as a state instead of a board.
For the time being this works to reduce code duplication.

Signed-off-by: Daniel R. Carvalho <odanrc@yahoo.com.br>