Fix: remove incorrectly named `ACPCState::Ante` function.
An ante refers to a forced bet that all players must make before receiving their cards. The `ACPCState::Ante` function instead returned the current number of chips a player had put into the pot. The function logic was actually an exact duplicate of the correctly named `ACPCState::CurrentSpent`, so we remove all instances of `ACPCState::Ante` and replace with that.

PiperOrigin-RevId: 804881589
Change-Id: I5f590858caa2b44413e34fc1133eb2b45055358e