Add MVP OpenSpiel tournament and cash game simulation for non-Holdem poker variants via RepeatedPokerkit
This allows OpenSpiel to support simulating (within a single game episode) multi-hand tournaments and cash games for the following poker variants:
- Razz
- SevenCardStud
- PotLimitOmaha
(in addition to the Limit and NoLimit hold'em variants it already could as of the last CL)

Also while here:

- fixed a bug in one of the asserts. Specifically, Pokerkit actually considers *effective* stack sizes, not absolute stack size, in fixed limit games when deciding to allow a player to bet less than the 'normal' minimum bet size for a street. (Meaning a player's bets can end up in that situation if all of the *other* remaining players in the hand have stacks that are too small.)

- made some of the 'VARIANT' lists/maps in pokerkit_wrapper non-private. Since it's very helpful for pokerkit users to be able to loop through all of the supported game variants + figure out what params it supports programmatically.

NOTE: This is only an MVP, see the TODOs for more details. In particular we plan to add a number of additional tests in the future.

NOTE: The 'returns' for the tournament's winner not containing all the stacks' chips  in some terminal states is an expected consequence of how we're currently handling player eliminations (which considers them to have busted out at **non-0 stacks sizes** in many cases). For more details see the warning in the code ending in "This can result in the total rewards not adding up to the number of total chips to start the tournament, so users should not rely on this assumption", + the corresponding logic in RepeatedPoker.

NOTE: This is a different game than the C++ 'RepeatedPoker' game, which uses universal_poker + a wrapped ACPC game (instead of pokerkit_wrapper + a wrapped pokerkit game). For more details see repeated_poker.cc.
PiperOrigin-RevId: 808837791
Change-Id: Ic2e578925c2f2ab53a023154dae25c80e6cf8269