Update hand history recording in `repeated_pokerkit`.
The `_wrapped_state_hand_histories` now stores the per-player hand histories obtained from `wrapped_state.to_struct().poker_hand_histories()` instead of a string representation of the entire wrapped state. This provides more detailed, user-friendly, and player-specific data while also using a _significantly_ smaller amount of bytes in total.

Additionally, I've added a bool to control whether we store these on the RepeatedPokerkit game itself. This way if we ever encounter this issue again, or if we just want to run an extremely large numbers of hands, we can simply flip the flag here to stop worrying about it. (As a result we did have to regenerate many of the playthroughs, however.)

PiperOrigin-RevId: 821063575
Change-Id: I4369154ff7d19b123ab0671ed12e3d9fba192338