Return Evaluate results from the perspective of player 0 rather than current player.
The Evaluator contract is that it returns estimated rewards for player 0
and 1, in that order. However, VPEvaluator would return {value, -value}
with value calculated based on ObservationTensor(), which is from a
perspective of current player, rather than player 0. Therefore, for half
of the states in the game, MCTS got the correct value, but with wrong
sign. This drastically affects the quality of chosen actions.